unified-monitor/requirements.txt
psutil
scapy
watchdog
python-systemd ; sys_platform == 'linux'
pywin32 ; sys_platform == 'win32'
prettytable
